Make gen-slt split cells based on type
There's a bigass dict that we can use to determine whether a particular
cell is an ice cave, a dwemer ruin, a civilized interior, etc.

These still have to be defined one by one, but once done, config for a
particular combination of plugin and interior type will be stored in a
folder named after that interior type and in a file named after that
plugin.

This should make it easier for users to find/replace their way to
creative freedom.